At leat, I'm starting to have something which looks useful.
The clone command (in pure-ocaml) now works:

    cagit clone <address> [--depth=<i>]

The first benchmarks seems to indicate that it is around ~3 time slower than the normal `git` command, but I haven't implemented any multicast options and I haven't looked very much at optimizing things, so lot of room for improvment here.

This just create a bare repostiory, need to unfold the filesystem from the recorded objects (but should be quite easy to do).

Also, just for fun, the second command of `cagit` does not exist in git:

    cagit grap -o FILE

This generate the graph of objects stored in the current repository. You need to be at the root of your repository to use it.

Note1: this will unpack all the pack objects, so you might want to run `git gc` after this to repack everything

Note2: this will work on any Git repository, no need to use `cagit clone` to use it