Do not skip deltas referecing shallow values
Shallow values are now properly translated to PIC values and the deltas
are not skipped anymore. The new delta objects points to the shallow PICs.
It's possible that two delta points to two different shallow PICs with
the same hash, that's annoying but should not be a bug as shallow PICs
are not serialized in the pack file anyway (so no risk of duplciation).