Better computation of haves when fetching
Now the client computes the "slice" of commits between the ones he knows about
and the ones the server knows about. This should optimize for most of the cases.

There is still a non-optimal case: when both the client and the server have
updated the same reference. In that case, we perform very badly. A possible
fix would be to send all the commits again in that case, but that's not very
elegant...