Use mmap for large files and read for small ones
Use of mmap is necessary to handle packfiles efficiently. Since these
are stored in a weak map, we won't run out of open files if we keep
accessing the same one.

Using read is necessary to handle references, since these are mutable
and can't be cached. Using mmap here leads to hitting the OS limit on
the number of open files.