Add Store.size
Before, the only way to get the size of a file was to decompress the
whole thing.

The main win comes from getting the size of raw packed objects without
decompressing. There are a couple of further improvements possible:

- Calculate the size of loose objects without loading them.

- For deltas in packs, only decompress the first few bytes (to get
  `result_size`).

This reduces the time to scan a large tree from 35s to 30s on my test
repository.