Add the new module Common
Common is a module which contains only some useful signatures like BASE,
ENCODER or DECODER. This module articulates all of this project.

For example, any Git object respects the signature BASE to provide a
pretty-printer, a compare function, an hash and an equal function. Then,
the interface require a specialized Set of the value t and a specialized
Map of the value t to play with the set theory with t -- like when we
compute a revision for example (e.g. abcdef..012345).

ENCODER and DECODER is to provide a non-blocking encoder/decoder with a
fixed size buffer. So, these module expect to read from an input with
[`Await] or write into an output with [`Flush].

This interface is close to an intf « à la bünzli » and I think it's more
relevant to have this intf than the idiom haskell (like [(input -> t)])
to keep a fine control about the memory consumption. The composition
between encoder/encoder, decoder/decoder is still possible - but needs
much more noise about the code.

From `ocaml-git`, it's the same than the module S.