Add the internal helper module.
This module is a toolbox about ocaml-git. It provides some useful
functor to make a specialized encoder or specialized decoder. The common
part of the hash implementation, an abstract digest function and a
function to save an OCaml value to a file from an Encoder.

For the commit 726fffc0, I talked about the composition of
encoder/encoder and decoder/decoder. Because the zlib interface is
constrained by INFLATE and DEFLATE (whose are close to ENCODER/DECODER),
we can compose any encoder with a deflater and any decoder with an
inflater. This is the purpose of MakeDeflater and MakeInflater.

NOTE: the inflater and the deflater is close to the same interface
provided by Decompress but I provide (some commits after) an
implementation with the same interface with camlzip.

Then, we have [digest] and [fdigest] whose make the digest function for
any object which has an encoder.

So, from an Angstrom decoder or from an Faraday/Farfadet encoder, you
can make an decocer which respects the intf Common.DECODER and an
encoder which respects the intf Common.ENCODER.

Finally, from an ENCODER, you can make a function to store an OCaml
value to a file.

This module is close to all about serialization/deserialization and it's
the big part of ocaml-git: how to serialize and deserialize a Git object
easily and fastly and keep a control about the memory consumption of
this kind of compute.

This module responds about this problematic in a good re-usable way pnly
for ocaml-git. For someone who looks this commit, don't copy/paste this
module. We have some assertion about what we try to decode (in this
case, a Git object) and this abstraction is possible only because we
know these assertions.