Add the minienc module
This module, close to Faraday, is here to respond to a specific
problematic about the memory consumption.

Indeed, Faraday was thinked to work with 2 process in the concurrency
way. One feed the Faraday encoder, the other one flush the content of
the encoder in the same order (a Queue).

However, in some kind of compute in ocaml-git, we don't follow this
pattern. In fact, this pattern is more reliable when you put a socket or
a file-descriptor as flusher. But, for [digest] function for example
(come from 0bf0e40), we put an DIGEST compute instead to produce an
hash.

And because, in this we can manipulate a big huge blob file for example,
it's needed to keep a fine control about the memory consumption of the
encoder and it's not the purpose of Faraday... But it's the purpose of
this module!

So, it provides an intf close to Faraday but provide the [`Flush] action
when the internal buffer is full. At this time, the client can flush the
content, feed something else (like a context for the hash), and continue
the process.