Add zlib implementation
As I said in 0bf0e40, the client can switch between the OCaml
implementation (decompress) and the C implementation (camlzip).

The OCaml implementation is just a specialization with Cstruct.

The C implementation is not thinked to be optimized but just to work.
So, may be a deep work is needed in these implemenations.

From my test, all seems to be ok but, because semantically [camlzip]
compute an input differently than [decompress], may be we can have some
problem. It's hard to find a common intf/semantic between these two
implementations. So, take care about these modules.

I did not do any benchmark. But, it seems to be easy to do one.