Add the Blob Git object
First module which represents a Git object. As I said in 726fffc, this
module respects the BASE intf. It provides a decoder (D), an encoder
(E), the Angstrom implementation (A), the Faraday implementation (F),
the minienc implementation (M) and provides the digest function (which
uses the minienc implementation).

I decide to define t as: private Cstruct.t to be easily for the end user
to manipulate a blob.