Add the synchronization module
This module is close to the old API about the fetching and the pushing
to a remote git repository but it uses the module provided by 017f626
commit, the Smart module.

This module takes care about the context shared by the client and the
server, specifically about the capabilities. Then, all action is
modulable to use an external computation (like external negociator,
external packer, etc.).