Clean unix implementation of the file-system and add an atomic read/write computation (@samoht)
This commit is little bit huge for some reasons:
* We handle the EINTR exception which re-call the current process
* We implement (as the previous version) the lock module (not finish!)
* We implement atomic read and write in the same way as the previous
  version

About the atomic read/write, we provide the function 'test_and_set' as
the previous version of the interface which will be used by the
reference module.

About the lock, one big difference is about the new way to handle the
serialization/deserialization. Previously, because all IO computation is
_atomic_, when the Git process try to write something, the lock exists
while the Git process does not finish to write/read all of the requested
file.

- Now, because I implemented all in the non-blocking way, I
cut in the Git process the open syscall and the read/write syscall.
These syscalls are protect by the locker (so we continue to read/write
sequentially) but we can have a conflict because nothing can stop the
Git process if it wants to read and (after) write __partially__ in the
same file.

So it's a partial commit for the file-system and the lock.