Fix the global lock about all mutable computation
Specifically, we put the lock abstraction in the reference
module, which represente the only mutable state of a Git repository.

To protect a race condition in a parallele/concurrent context, we
protect the write/test_and_set (read ?) computation by a global lock.
However, (@samoht), by the design (non-blocking computation), the write
function need to know the abstraction of the Lock implementation to use
it precisely as it uses the open/write syscall (see 199e3ff).

We can avoid that if we request in the file-system implementation an
atomic write (already available for [test_and_set] but not exposed) and
lint what is needed to write a reference (and protect it).

This commit follows the same semantic from the previous version but it
not tested yet.