Add flush function queue in minienc to apply a side-effect function as a renderer
This feature has the same purpose than Faraday.flush. You can record a
function at a specific point and when it's a time to flush some buffers,
if we traverse the point, we call __before__ the recorded function.

It's useful for the pkt-line and fill the buffer when we complete the
current line.