Delete the specialized implementation of some modules
Before this commit, in the Git directory, you can find some useful
implementation to apply the Store/Mem functor and test it in the
Unix/Jsoo environment specifically.

However, the purpose of the git directory is to contain only the Git
implementation indenpendantly of the back-end (Unix, Mirage, JS). The
code deleted was not lost, it just moved to another directories.