Move the interface of Ihash to the S module.
The first task is: rename [Ihash] to [Hash] and use then [Hash.S] as
the interface of the hash module

However, if you rename [Ihash] to [Hash.S], we have a problem with
this line:

> include Hash.DIGEST with type t := t and type hash = Hash.t

It's a clash of name of the name of the module [Hash] declared in a
previous line:

> module Hash : Hash.S

So, the better way is to move the interface (because it's only an
interface - no implementation provided) to the S module which contains
all other interfaces (like ENCODER/DECODER). We centralize all
definition in one module finally (and the next commit has the same
purpose).