Delete fs.mli and move the interface to the mode S
We add a new type +'a io in the interface and constraint it by:

'a File.io = 'a io
'a Dir.io = 'a io
'a Mapper.io = 'a io

That means the implementation of the file-system use only Lwt xor
Async (not both). Then, it's about the Lock implementation but you
can see the commit d581995.