Provided constraints to prove the equality of the Value.t between any applied functor
The goal of this large commit is to provide a description at each step
of the API to know exactly where come from the Value.t and ensure a
type equality between 2 implementations which provide the Value.t.

So, if a module A provide Value.t which created with a module Hash, we
can prove the equality with B.Value.t if B created the module Value
with the same module Hash.

Then, we can move a value from a memory back-end to a file-system (and
vice-versa) without any casting.