Add a Sync interface
This commit is little big huge. The main goal is to uniformism the
interface between sync and sync_http module - obviously, not at all,
like the ~header argument not relevant for the tcp implementation of
the smart protocol.

We put some log informations too and delete useless code (use instead
the Pack module provided by the Store implementation).