Fixed the engine behind the Smart protocol on the HTTP overlay when we fetch
In this patch, we consider (than before) 2 cases at the beginning:

- first, if the current client has nothing (clone), in this special
  case, it does not need to abstract something: we just need to send
  DONE and process the negotiation result.

- second, if the current client has something. Then, we need to start
  a negotiation with the server (this negotiation starts in any
  sub-case with what the client has and a FLUSH packet)

Then, it's essentially what is going on the server side and the client
side. The common case is to continue the negotiation because neither
the server responded by `Ready nor the client (the negotiation engine)
responded by `Done. This case is `Again.

If the server responded by `Ready, we expect a negotiation result
directly then - and start to unpack. This case appear only if both
server and client has the `Multi_ack_detailed capability (but this
verification is already donei before).

If the client responded by `Done, we need to /canonicalize/ what the
server already responded by ACK with what we already send. This
strongly common list between the client and the server will be send
/again/ and because the server already responded by ACK for all of
these hashes, we expect, again, than the server ACK __all__ (and not
miss one of them).

But, and because the client is `Done, instead to send a `Flush at the
end, we send a `Done packet. The server will ACK all and will send a
negotiation response.

Now, this is why we change the sematic of [Smart.p_negociation].
Because we send __only__ common hashes between the client and the
server, the server will ACK all. Then, if we follow the new semantic
of [Smart.p_negociation], this parser will end up not at the end of
the input but when it processed all ACKs.

Finally, in the input (because we send the `Done packet), the server
send the negotiation result and we can process it with
[negotiation_result] and finally unpack.