Use with_open instead of explicit open/close
This clean-up a lot of code and remove deduplication of handling
errors in case close is failing (or not).

Also there are a few suspicious bits in the previous code where it is not
totally clear that all open are properly closed. I let one pair of open/close
in pack_engine as the surrounding code is a bit complex to untangle. I will
probably fix that later.