Clean git-unix module entry--point
In this commit, we use Store.Make instead Store.FS first. Indeed, it's
redundant to have an Fs module (which is a really implementation of a
File-System) and the FS (from Store.FS) module which does not mean a
relevant thing.

Then, we provide a function with_fs which let the client to make a store
from an user defined file-system (so outside the scope of ocaml-git).
The, the INDEX implementation AND the store can play with the same fs
value needed to interact with the file-system.

However, INDEX takes some liberties to use directly the unix module for
some "special" syscall which are not available on the file-system
abstraction.