Fix git-mirage backend
We enforce to flush `Channel` when we write something to ensure to
expect something to read then.

We update the API Git.Sync.NET to let I/O implementation to return
an error and raise it to the end user ([`Net] error).

Finally, we raise correctly a parsing error to the end-user ([`Smart]
error)