Add Stream module
This module is a replacement of old [Helper] module. It provides a way to
calculate the hash of a Git object given by its OCaml representation.

The way to calculate the hash is:
1) serialise the Git object
2) start with a /header/ ([kind length\000])
3) feed the context with the serialised Git object

A Git object can be big. Instead to entirely serialise it, we /stream/ it to
limit the memory footprint. NOTE: the memory footprint is __not__ really
limited (see [test/tree/test] to understand why) but, at least, the serialised
value is cut to many /small/ [string].