Provide new I/O operations on References from a simple [store] impl.
This patch follows the same global design, a /de-functorization/ of Git.
Git Reference need a simple store with few operations. All of these can fit
into a simple record (with the help of the high kind polymorphism provided by
[Carton]).

Finally, [Mem] and [Store] will specialise it with LWT according to an
[Hashtbl] (for [Mem]) or a given implementation of an /atomic/ store (for
[Store]).