Integrate Not-So-Smart
Not So Smart (nss) is an implementation of the Smart protocol needed to push
and fetch to a Git repository. This new implementation does not require a
strong dependency to Git and can exist from a small set of accessors of a Git
repository.

This set can be implementation with a direct call to [git] (as it is done by
tests) or with the given implementation of Git described by the minimal
interface of a Git store.

By design, it separates the protocol from the store implementation. Only a way
to manipulate/generate a PACK file is required and provided by [Carton].