Gri module was /conceptually/ move into the nss library
The representation of the endpoint was moved into [nss] as [Smart_git.endpoint]
which gives to us a better representation (including SSH endpoint) than the
current one to dispatch the flow to /the right implementation/.

The /right implementation/ is a common impl. of the Smart protocol but it
integrate few details depending on the given endpoint - see [nss] for more
details.