TCP module was deleted/moved into the new nss library
Due to [nss], the difference between the TCP, the HTTP or the SSH communication
to a Git remote point does not exist anymore. A common way to communicate to
these endpoints and [nss] ensure to tweak details to be compatible with Git.

So [TCP] has no reason to exist anymore.