Add an implementation of /loose/ files
In the way to redesign the core library, a new sub-library exists to implement
the logic to extract/add a Git loose object into an append-only store. It
ensures to deflate/inflate correctly objects and it requires a specific
layout from the append-only space, close to a radix-tree (where ['uid] used to
identify a Git object is the hash of the object).

It will replace the old implementation of loose file implemented into the core
library of [ocaml-git].