Smart module was /conceptually/ moved into the new nss library
The implementation of the Smart protocol without the logic of the Smart
protocol was moved into [nss] as [protocol.ml] which is a better implementation
of it (it fixes few bugs about the PKT-format).

The new implementation gives to us more possibilities about abstractions - see
the [nss] library about that.