Integrate Carton
Carton is an implementation to encode/decode a PACKv2 file. This implementation
does not depend on any /syscalls/ and any /schedulers/ - so it can be used by
MirageOS or UNIX env.

The implementation provides a fast way to extract an object. However, Carton
does not have the logic of a Git object - it can store /any/ objects. The
library is able to produce a PACKv2 file too from a list of objects.

The implementation unlocks the ability for us to test the implementation
independently of Git. This implementation unlocks the ability to implement
Not-So-Smart without an Git OCaml implementation.