Delete the way to handle PACK file (from disk or network).
Current implementation is too complex to handle PACK files which come from the
disk (with an IDX file) or the network (which can be /thin/). All of these
will be replaced by a smaller/smarter implementation where the big analyse is
done by [Carton.Thin] which is able to /re-canonicalize/ a /thin/ PACK file.