[git-mirage] is not needed anymore where it is a specialisation of [Sync] over Conduit
The new version of Conduit is small enough to depend on it and use it. At least,
the core Git library requires the core Conduit library. The old [git-mirage]
package was just a specialisation about /which context we use to start a
connection/ but Conduit itself is enough to get this kind of abstr. - and
Conduit was thought of as such.